Create AttributesSerializer with get_attribute and to_representation methods
